pyproject.toml
[project]
name = "elastic-query-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elasticsearch==8.11.1",
    "pydantic-ai>=0.4.0",
    "pydantic>=2.11.7",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "fastapi>=0.116.0",
    "pandas>=2.3.1",
    "openpyxl>=3.1.5",
    "streamlit>=1.46.1",
    "graphviz>=0.21",
    "pymongo>=4.13.2",
]
